internals

internal documentation

Documentation for PhyloPlots's internal functions. Those functions are not exported, but can still be used (like: PhyloPlots.foo() for a function named foo()).

<!–@contents Pages = ["internals.md"]–>

index

PhyloPlots.checkNodeDataFrame

PhyloPlots.getEdgeNodeCoordinates

PhyloPlots.prepareEdgeDataFrame

PhyloPlots.prepareNodeDataFrame

PhyloPlots.rexport

<!– ## Types –>

functions

PhyloPlots.checkNodeDataFrame — Method.

checkNodeDataFrame(net, nodeLabel)

Check data frame for node annotations:

check that the data has at least 2 columns (if it has any)

check that the first column has integers (to serve as node numbers)

remove rows with no node numbers

warning if some node numbers in the data are not in the network.

source

PhyloPlots.getEdgeNodeCoordinates — Method.

getEdgeNodeCoordinates(net, useEdgeLength)

Calculate coordinates for plotting later with Gadfly or RCall.

Actually modifies some (minor) attributes of the network, as it calls directEdges!, preorder! and cladewiseorder!.

source

PhyloPlots.prepareEdgeDataFrame — Method.

prepareEdgeDataFrame(net, edgeLabel::DataFrame, mainTree::Bool,
    edge_xB, edge_xE, edge_yB, edge_yE)

Check data frame for edge annotation. edge_*: Float64 vectors giving the coordinates for the beginning and end of edges. Return data frame with columns

x, y: coordinates on the plots

len: node name

gam: gamma (inheritance value)

num: node number

lab: node label

hyb: is hybrid?

min: is minor?

source

PhyloPlots.prepareNodeDataFrame — Method.

prepareNodeDataFrame(net, nodeLabel::DataFrame,
    showNodeNumber::Bool, showIntNodeLabel::Bool, labelnodes::Bool,
    node_x, node_y)

Make data frame for node annotation. node_* should be Float64 vectors. nodeLabel should have columns as required by checkNodeDataFrame

Columns of output data frame:

x, y: coordinates on the plots (from node_*)

name: node name

num: node number

lab: node label

lea: is leaf?

source

PhyloPlots.rexport — Method.

rexport(net::HybridNetwork; mainTree=false, useEdgeLength=true)

Create an RObject of class phylo (and evonet depending on the number of hybridizations) recognized by the ape library in R (S3 object). This RObject can be evaluated using the tools available in the ape library in R. For example, we can visualize the network using ape's plot function.

not exported: sexp is the best way to go.

Arguments

useEdgeLength: if true, export edge lengths from net.

mainTree: if true, minor hybrid edges are omitted.

Examples

julia> net = readTopology("(((A,(B)#H1:::0.9),(C,#H1:::0.1)),D);");
julia> phy = rexport(net)
RCall.RObject{RCall.VecSxp}
$Nnode
[1] 5

$edge
     [,1] [,2]
[1,]    5    6
[2,]    5    4
[3,]    6    8
[4,]    6    7
[5,]    7    3
[6,]    8    1
[7,]    8    9
[8,]    9    2

$tip.label
[1] "A" "B" "C" "D"

$reticulation
     [,1] [,2]
[1,]    7    9

$reticulation.gamma
[1] 0.1

attr(,"class")
[1] "evonet" "phylo"

julia> using RCall

julia> R"library(ape)";

julia> phy
RCall.RObject{RCall.VecSxp}

    Evolutionary network with 1 reticulation

               --- Base tree ---
Phylogenetic tree with 4 tips and 5 internal nodes.

Tip labels:
[1] "A" "B" "C" "D"

Rooted; no branch lengths.

R> phy

Evolutionary network with 1 reticulation

               --- Base tree ---
Phylogenetic tree with 4 tips and 5 internal nodes.

Tip labels:
[1] "A" "B" "C" "D"

Rooted; no branch lengths.

R> str(phy)
List of 5
$ Nnode             : int 5
$ edge              : int [1:8, 1:2] 5 5 6 6 7 8 8 9 6 4 ...
$ tip.label         : chr [1:4] "A" "B" "C" "D"
$ reticulation      : int [1, 1:2] 7 9
$ reticulation.gamma: num 0.1
- attr(*, "class")= chr [1:2] "evonet" "phylo"

source