public documentation

Documentation for PhyloPlots's public (exported) functions.

index

Gadfly.plot

Gadfly.plot

Gadfly.plot

RCall.sexp

functions

Gadfly.plot — Method

plot(net::HybridNetwork, method::Symbol)

Plot a network using R graphics. method should be :R (actually, any symbol would do, for now!).

optional arguments, shared with the Gadfly-based plot function:

useEdgeLength = false : if true, the tree edges and major hybrid edges are drawn proportionally to their length. Minor hybrid edges are not, however. Note that edge lengths in coalescent units may scale very poorly with time.

showTipLabel = true : if true, taxon labels are shown. You may need to zoom out to see them.

showNodeNumber = false : if true, nodes are labelled with the number used internally.

showEdgeLength = false : if true, edges are labelled with their length (above)

showGamma = false : if true, hybrid edges are labelled with their heritability (below)

edgeColor = "black" : color for tree edges.

majorHybridEdgeColor = "deepskyblue4" : color for major hybrid edges

minorHybridEdgeColor = "deepskyblue" : color for minor hybrid edges

showEdgeNumber = false : if true, edges are labelled with the number used internally.

showIntNodeLabel = false : if true, internal nodes are labelled with their names. Useful for hybrid nodes, which do have tags like 'H1'.

edgeLabel = DataFrame() : dataframe with two columns: the first with edge numbers, the second with labels (like bootstrap values) to annotate edges. empty by default.

nodeLabel = DataFrame() : dataframe with two columns: the first with node numbers, the second with labels (like bootstrap values for hybrid relationships) to annotate nodes. empty by default.

style = :fulltree : symbol indicating the style of the diagram

:majortree will simply draw minor edges onto the major tree.

:fulltree will draw minor edges as their own branches in the tree (like in icytree.org), useful for overlapping or confusing networks.

arrowlen : the length of the arrow tips in the full tree style. if style = :fulltree, then arrowlen = 0.2. otherwise, arrowlen = 0, which makes the arrows appear as segments.

optional arguments specific to this function:

xlim, ylim : array of 2 values

tipOffset = 0.0 : to offset tip labels

plot() returns the following tuple: (xmin, xmax, ymin, ymax, node_x, node_y, node_yB, node_yE, edge_xB, edge_xE, edge_yB, edge_yE, ndf, edf)

xmin : the minimum x value of the plot

xmax : the maximum x value of the plot

ymin : the minimum y value of the plot

ymax : the maximum y value of the plot

node_x : the x values of the nodes in net.node in their respective order

node_y : the y values of the nodes

node_yB : the y value of the beginning of the verticle bar

node_yE : the y value of the end of the verticle bar

node_yE : the y value of the end of the verticle bar

edge_xB : the x value of the beginning of the edges in net.edge in their respective order

edge_xE : the x value of the end of the edges

edge_yB : the y value of the beginning of the edges

edge_yE : the y value of the end of the edges

ndf : the node data frame: see section Adding labels for more

edf : the edge data frame

Note that plot actually modifies some (minor) attributes of the network, as it calls directEdges! and preorder!.

If hybrid edges cross tree and major edges, you may choose to rotate some tree edges to eliminate crossing edges, using rotate! (in PhyloNetworks).

Alternative: a tree or network can be exported with sexp and then displayed with R's "plot" and all its options.

Gadfly.plot — Method

plot(net::HybridNetwork; useEdgeLength=false, mainTree=false, showTipLabel=true,
     showNodeNumber=false, showEdgeLength=false, showGamma=false, edgeColor=colorant"black",
     majorHybridEdgeColor=colorant"deepskyblue4", minorHybridEdgeColor=colorant"deepskyblue",
     showEdgeNumber=false, showIntNodeLabel=false, edgeLabel=[], nodeLabel=[])

Plots a network, from left to right.

useEdgeLength: if true, the tree edges and major hybrid edges are drawn proportionally to their length. Minor hybrid edges are not, however. Note that edge lengths in coalescent units may scale very poorly with time.

mainTree: if true, the minor hybrid edges are ommitted.

showTipLabel: if true, taxon labels are shown. You may need to zoom out to see them.

showNodeNumber: if true, nodes are labelled with the number used internally.

showEdgeLength: if true, edges are labelled with their length (above)

showGamma: if true, hybrid edges are labelled with their heritability (below)

edgeColor: color for tree edges. black by default.

majorHybridEdgeColor: color for major hybrid edges

minorHybridEdgeColor: color for minor hybrid edges

showEdgeNumber: if true, edges are labelled with the number used internally.

showIntNodeLabel: if true, internal nodes are labelled with their names. Useful for hybrid nodes, which do have tags like 'H1'.

edgeLabel: dataframe with two columns: the first with edge numbers, the second with labels (like bootstrap values) to annotate edges. empty by default.

nodeLabel: dataframe with two columns: the first with node numbers, the second with labels (like bootstrap values for hybrid relationships) to annotate nodes. empty by default.

Note that plot actually modifies some (minor) attributes of the network, as it calls directEdges!, preorder! and cladewiseorder!.

If hybrid edges cross tree and major edges, you may choose to rotate some tree edges to eliminate crossing edges, using rotate!.

Gadfly.plot — Method

plot(model::TwoBinaryTraitSubstitutionModel)

Display substitution rates for a trait evolution model for two possibly dependent binary traits; using R via RCall. Adapted from fitPagel functions found in the R package phytools.

Examples

julia> using PhyloNetworks

julia> m = TwoBinaryTraitSubstitutionModel([2.0,1.2,1.1,2.2,1.0,3.1,2.0,1.1],
                ["carnivory", "noncarnivory", "wet", "dry"])
Substitution model for 2 binary traits, with rate matrix:
                     carnivory-wet    carnivory-dry noncarnivory-wet noncarnivory-dry
    carnivory-wet                *           1.0000           2.0000           0.0000
    carnivory-dry           3.1000                *           0.0000           1.1000
 noncarnivory-wet           1.2000           0.0000                *           2.0000
 noncarnivory-dry           0.0000           2.2000           1.1000                *

julia> plot(m);

RCall.sexp — Method

function sexp(net::HybridNetwork)

Export a HybridNework object to the R language as either phylo or evonet object (depending on degree of hybridization) recognized by the R package ape. Used by the $object syntax and by @rput to use a Julia object in R: see the examples below. Makes it easy to plot a Julia tree or network using plotting facilities in R.

code inspired from Phylo.jl

Examples

julia> using RCall
julia> using PhyloNetworks
julia> net = readTopology("(((A:.2,(B:.1)#H1:.1::0.9):.1,(C:.11,#H1:.01::0.1):.19):.1,D:.4);");
R> library(ape); # type $ to switch from julia to R
R> $net

Evolutionary network with 1 reticulation

               --- Base tree ---
Phylogenetic tree with 4 tips and 5 internal nodes.

Tip labels:
[1] "A" "B" "C" "D"

Rooted; includes branch lengths.
julia> @rput net # press the delete key to switch from R back to julia
R> net

Evolutionary network with 1 reticulation

--- Base tree ---
Phylogenetic tree with 4 tips and 5 internal nodes.

Tip labels:
[1] "A" "B" "C" "D"

Rooted; includes branch lengths.

R> str(net)
List of 7
 $ edge               : int [1:8, 1:2] 5 5 6 6 7 8 8 9 6 4 ...
 $ reticulation.length: num 0.01
 $ Nnode              : int 5
 $ edge.length        : num [1:8] 0.1 0.4 0.1 0.19 0.11 0.2 0.1 0.1
 $ reticulation       : int [1, 1:2] 7 9
 $ reticulation.gamma : num 0.1
 $ tip.label          : chr [1:4] "A" "B" "C" "D"
 - attr(*, "class")= chr [1:2] "evonet" "phylo"
NULL

R> plot(net)