public documentation

Documentation for PhyloPlots's public (exported) functions.

index

Gadfly.plot

Gadfly.plot

Gadfly.plot

RCall.sexp

functions

Gadfly.plot — Method

plot(net::HybridNetwork, method::Symbol)

Plot a network using R graphics. method should be :R (actually, any symbol would do, for now!).

optional arguments, shared with the Gadfly-based plot function:

useEdgeLength: if true, the tree edges and major hybrid edges are drawn proportionally to their length. Minor hybrid edges are not, however. Note that edge lengths in coalescent units may scale very poorly with time.

showTipLabel: if true, taxon labels are shown. You may need to zoom out to see them.

showNodeNumber: if true, nodes are labelled with the number used internally.

showEdgeLength: if true, edges are labelled with their length (above)

showGamma: if true, hybrid edges are labelled with their heritability (below)

edgeColor: color for tree edges. black by default.

majorHybridEdgeColor: color for major hybrid edges

minorHybridEdgeColor: color for minor hybrid edges

showEdgeNumber: if true, edges are labelled with the number used internally.

showIntNodeLabel: if true, internal nodes are labelled with their names. Useful for hybrid nodes, which do have tags like 'H1'.

edgeLabel: dataframe with two columns: the first with edge numbers, the second with labels (like bootstrap values) to annotate edges. empty by default.

nodeLabel: dataframe with two columns: the first with node numbers, the second with labels (like bootstrap values for hybrid relationships) to annotate nodes. empty by default.

optional arguments specific to this function:

xlim, ylim: array of 2 values

tipOffset: to offset tip labels

Note that plot actually modifies some (minor) attributes of the network, as it calls directEdges!, preorder! and cladewiseorder!.

If hybrid edges cross tree and major edges, you may choose to rotate some tree edges to eliminate crossing edges, using rotate! (in PhyloNetworks).

Alternative: a tree or network can be exported with sexp and then displayed with R's "plot" and all its options.

Gadfly.plot — Method

plot(net::HybridNetwork; useEdgeLength=false, mainTree=false, showTipLabel=true,
     showNodeNumber=false, showEdgeLength=false, showGamma=false, edgeColor=colorant"black",
     majorHybridEdgeColor=colorant"deepskyblue4", minorHybridEdgeColor=colorant"deepskyblue",
     showEdgeNumber=false, showIntNodeLabel=false, edgeLabel=[], nodeLabel=[])

Plots a network, from left to right.

useEdgeLength: if true, the tree edges and major hybrid edges are drawn proportionally to their length. Minor hybrid edges are not, however. Note that edge lengths in coalescent units may scale very poorly with time.

mainTree: if true, the minor hybrid edges are ommitted.

showTipLabel: if true, taxon labels are shown. You may need to zoom out to see them.

showNodeNumber: if true, nodes are labelled with the number used internally.

showEdgeLength: if true, edges are labelled with their length (above)

showGamma: if true, hybrid edges are labelled with their heritability (below)

edgeColor: color for tree edges. black by default.

majorHybridEdgeColor: color for major hybrid edges

minorHybridEdgeColor: color for minor hybrid edges

showEdgeNumber: if true, edges are labelled with the number used internally.

showIntNodeLabel: if true, internal nodes are labelled with their names. Useful for hybrid nodes, which do have tags like 'H1'.

edgeLabel: dataframe with two columns: the first with edge numbers, the second with labels (like bootstrap values) to annotate edges. empty by default.

nodeLabel: dataframe with two columns: the first with node numbers, the second with labels (like bootstrap values for hybrid relationships) to annotate nodes. empty by default.

Note that plot actually modifies some (minor) attributes of the network, as it calls directEdges!, preorder! and cladewiseorder!.

If hybrid edges cross tree and major edges, you may choose to rotate some tree edges to eliminate crossing edges, using rotate!.

Gadfly.plot — Method

plot(model::TwoBinaryTraitSubstitutionModel)

Display substitution rates for a trait evolution model for two possibly dependent binary traits; using R via RCall. Adapted from fitPagel functions found in the R package phytools.

Examples

julia> using PhyloNetworks

julia> m = TwoBinaryTraitSubstitutionModel([2.0,1.2,1.1,2.2,1.0,3.1,2.0,1.1],
                ["carnivory", "noncarnivory", "wet", "dry"])
Substitution model for 2 binary traits, with rate matrix:
                     carnivory-wet    carnivory-dry noncarnivory-wet noncarnivory-dry
    carnivory-wet                *           1.0000           2.0000           0.0000
    carnivory-dry           3.1000                *           0.0000           1.1000
 noncarnivory-wet           1.2000           0.0000                *           2.0000
 noncarnivory-dry           0.0000           2.2000           1.1000                *

julia> plot(m);

RCall.sexp — Method

function sexp(net::HybridNetwork)

Export a HybridNework object to the R language as either phylo or evonet object (depending on degree of hybridization) recognized by the R package ape. Used by the $object syntax and by @rput to use a Julia object in R: see the examples below. Makes it easy to plot a Julia tree or network using plotting facilities in R.

code inspired from Phylo.jl

Examples

julia> using RCall
julia> using PhyloNetworks
julia> net = readTopology("(((A:.2,(B:.1)#H1:.1::0.9):.1,(C:.11,#H1:.01::0.1):.19):.1,D:.4);");
R> library(ape); # type $ to switch from julia to R
R> $net

Evolutionary network with 1 reticulation

               --- Base tree ---
Phylogenetic tree with 4 tips and 5 internal nodes.

Tip labels:
[1] "A" "B" "C" "D"

Rooted; includes branch lengths.
julia> @rput net # press the delete key to switch from R back to julia
R> net

Evolutionary network with 1 reticulation

--- Base tree ---
Phylogenetic tree with 4 tips and 5 internal nodes.

Tip labels:
[1] "A" "B" "C" "D"

Rooted; includes branch lengths.

R> str(net)
List of 7
 $ edge               : int [1:8, 1:2] 5 5 6 6 7 8 8 9 6 4 ...
 $ reticulation.length: num 0.01
 $ Nnode              : int 5
 $ edge.length        : num [1:8] 0.1 0.4 0.1 0.19 0.11 0.2 0.1 0.1
 $ reticulation       : int [1, 1:2] 7 9
 $ reticulation.gamma : num 0.1
 $ tip.label          : chr [1:4] "A" "B" "C" "D"
 - attr(*, "class")= chr [1:2] "evonet" "phylo"
NULL

R> plot(net)